Small object Ids refactor and extract external repository use cases

+++ src/main/java/io/pakland/mdas/githubstats/application/mappers/UserMapper.java
@@ -5,6 +5,6 @@ import io.pakland.mdas.githubstats.domain.User;
 
 public class UserMapper {
     public static User dtoToEntity(UserDTO dto) {
-        return User.builder().id(dto.getId().longValue()).login(dto.getLogin()).build();
+        return User.builder().id(dto.getId()).login(dto.getLogin()).build();
     }
 }
